+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 504 Add Elasticsearch module (@chertby)
 - 516 Add `ITestcontainersBuilder<TDockerContainer>.WithTmpfsMount` (@chrisbbe)
 - 520 Add MariaDB module (@renemadsen)
+- 528 Do not require the Docker host configuration (`DockerEndpointAuthConfig`) on `TestcontainersSettings` initialization
 
 ### Fixed